Unify tracing and metrics export.
