+++ fuzzysearch-common/src/lib.rs
@@ -4,6 +4,9 @@ pub mod types;
 #[cfg(feature = "video")]
 pub mod video;
 
+#[cfg(feature = "trace")]
+pub mod trace;
+
 /// Create an instance of img_hash with project defaults.
 pub fn get_hasher() -> img_hash::Hasher<[u8; 8]> {
     use img_hash::{HashAlg::Gradient, HasherConfig};
