+++ fuzzysearch-ingest-e621/Cargo.toml
@@ -16,7 +16,6 @@ serde_json = "1"
 sqlx = { version = "0.5", features = ["runtime-tokio-native-tls", "postgres", "macros", "json", "offline"] }
 
 image = "0.23"
-hamming = "0.1.3"
 img_hash = "3"
 sha2 = "0.9"
