+++ fuzzysearch-ingest-furaffinity/Cargo.toml
@@ -6,11 +6,8 @@ edition = "2018"
 
 [dependencies]
 reqwest = "0.11"
-postgres = { version = "0.19", features = ["with-chrono-0_4"] }
 tokio = { version = "1", features = ["full"] }
-tokio-postgres = { version = "0.7.0" }
-r2d2_postgres = " 0.18.0"
-r2d2 = "0.8"
+tokio-postgres = { version = "0.7.0", features = ["with-chrono-0_4"] }
 chrono = "0.4"
 hyper = { version = "0.14", features = ["server"] }
 prometheus = { version = "0.12", features = ["process"] }
