+++ fuzzysearch-ingest-weasyl/src/main.rs
@@ -198,7 +198,8 @@ async fn insert_null(
 
 #[tokio::main]
 async fn main() {
-    fuzzysearch_common::init_logger();
+    fuzzysearch_common::trace::configure_tracing();
+    fuzzysearch_common::trace::serve_metrics().await;
 
     let api_key = std::env::var("WEASYL_APIKEY").unwrap_or_log();
