+++ fuzzysearch-webhook/src/main.rs
@@ -27,7 +27,7 @@ pub enum WebhookError {
 }
 
 fn main() {
-    fuzzysearch_common::init_logger();
+    fuzzysearch_common::trace::configure_tracing();
 
     tracing::info!("Starting...");
