+++ fuzzysearch/Cargo.toml
@@ -18,7 +18,6 @@ tracing-opentelemetry = "0.12"
 opentelemetry-http = "0.2"
 
 tokio = { version = "1", features = ["full"] }
-async-stream = "0.3"
 
 futures = "0.3"
